+++ netbox_fms/__init__.py
@@ -20,9 +20,10 @@ class NetBoxFMSConfig(PluginConfig):
 
     def ready(self):
         super().ready()
-        from netbox_fms.monkey_patches import patch_cable_profiles
+        from netbox_fms.monkey_patches import patch_cable_profiles, patch_delete_origin
 
         patch_cable_profiles()
+        patch_delete_origin()
         from .signals import connect_signals
 
         connect_signals()